+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
- * (Nothing yet)
+ * Add ability to write interpolated Trial ensemble with `ensPostProcess` (#425 and !405)
 
 ### Changed